Reward MVPs on each team
This isn't perfect since players can leave and join other teams and
"take over" another team's mvp slots, but this is ok for now.

We can mitigate this in the future by not allowing players to rejoin a
warzone on a different team or by having more granular stat tracking.